Adjusting the unixTimestampToDate function and its test
