+++ apps/ensadmin/src/components/recent-registrations/utils.tsx
@@ -1,5 +1,6 @@
 import { formatDistanceStrict, formatDistanceToNow, intlFormat } from "date-fns";
 import { useEffect, useState } from "react";
+import {millisecondsInSecond} from "date-fns/constants";
 
 /**
  * Client-only date formatter component
@@ -50,4 +51,22 @@ export function Duration({
   }, [beginsAt, endsAt]);
 
   return <>{duration}</>;
+}
+
+/**
+ * An integer value (representing a Unix timestamp in seconds) formatted as a string.
+ */
+export type UnixTimestampInSeconds = string;
+
+/**
+ * Transforms a timestamp formatted as a string to a Date object.
+ */
+export function unixTimestampToDate(timestamp: UnixTimestampInSeconds): Date {
+  const date = new Date(parseInt(timestamp) * millisecondsInSecond);
+
+  if (isNaN(date.getTime())) {
+    throw new Error(`Error parsing timestamp (${timestamp}) to date`);
+  }
+
+  return date;
 }
