+++ apps/ensadmin/tsconfig.json
@@ -21,7 +21,6 @@
     "paths": {
       "@/*": ["./src/*"]
     },
-    "types": ["vitest/importMeta"]
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
   "exclude": ["node_modules"]
