+++ apps/ensadmin/vitest.config.ts
@@ -2,12 +2,6 @@ import { resolve } from "path";
 import { defineConfig } from "vitest/config";
 
 export default defineConfig({
-  test: {
-    includeSource: ["src/**/*.{js,ts}"],
-  },
-  define: {
-    "import.meta.vitest": "undefined",
-  },
   resolve: {
     alias: {
       "@": resolve(__dirname, "./src"),
